Bug #1347: Add support for float quotas. The update quotas method needs to be changed too (cherry picked from commit 6cc05f5b81c7cdcf302d255a3d6490ea87921cc1)

+++ src/um/QuotaDatastore.cc
@@ -28,7 +28,7 @@ const int QuotaDatastore::NUM_DS_METRICS  = 2;
 
 bool QuotaDatastore::check(Template * tmpl,  string& error)
 {
-    map<string, int> ds_request;
+    map<string, float> ds_request;
 
     string ds_id;
     int    size;
